requirements.txt
Flask==3.1.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
numpy==1.24.4
pillow==11.0.0
scikit-learn==1.5.2
scipy==1.13.1
torch==2.2.2
torchvision==0.17.2
